+++ test/Veggerby.Boards.Tests/Integration/CompiledPatterns/ChessCompiledIntegrationParityTests.cs
@@ -1,8 +1,9 @@
 using System.Linq;
+
+using Veggerby.Boards.Artifacts.Relations;
 using Veggerby.Boards.Chess;
 using Veggerby.Boards.Internal;
 using Veggerby.Boards.States;
-using Veggerby.Boards.Artifacts.Relations;
 
 using Xunit;